+++ ex03/includes/AForm.hpp
@@ -10,20 +10,20 @@ class AForm
 private:
   const std::string _name;
   bool _signed;
-  const size_t _signGrade;
-  const size_t _exeGrade;
+  const int _signGrade;
+  const int _exeGrade;
 
 public:
   AForm();
   AForm(const AForm &src);
-  AForm(std::string name, size_t signGrade, size_t exeGrade);
+  AForm(std::string name, int signGrade, int exeGrade);
   virtual ~AForm();
 
   AForm &operator=(const AForm &rhs);
 
   std::string const &getName(void) const;
-  size_t const &getSignGrade(void) const;
-  size_t const &getExeGrade(void) const;
+  int const &getSignGrade(void) const;
+  int const &getExeGrade(void) const;
   bool const &getSignedState(void) const;
 
   void beSigned(Bureaucrat &b);
